Extract parameter query tests to separate file.
